refactor(queries): consolidate alt text steps into a single file
