+++ src/queries/index.ts
@@ -1,5 +1,4 @@
 export * from './alt-text';
-export * from './alt-texts';
 export * from './button';
 export * from './buttons';
 export * from './children';
